[feat] postId, userId의 타입을 int에서 long으로 변경

+++ src/main/java/me/sisyphusj/community/app/auth/service/CustomOAuth2UserService.java
@@ -77,7 +77,7 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
 	/**
 	 * 기존 가입자인지 확인 <br> 최초 로그인 시 가입 절차 진행
 	 */
-	private int saveOrUpdate(OAuthDTO oAuthDTO) {
+	private long saveOrUpdate(OAuthDTO oAuthDTO) {
 		Optional<OAuthVO> oAuthVO = authMapper.selectOAuthByUsername(oAuthDTO.getUsername());
 		return oAuthVO.map(OAuthVO::getUserId)
 			.orElseGet(() -> signupOAuth(oAuthDTO));
@@ -86,7 +86,7 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
 	/**
 	 * 가입과 동시에 반환된 oAuthVO 객체로부터 userId 반환
 	 */
-	private int signupOAuth(OAuthDTO oAuthDTO) {
+	private long signupOAuth(OAuthDTO oAuthDTO) {
 		OAuthVO oAuthVO = OAuthVO.of(oAuthDTO);
 		authMapper.insertOAuth(oAuthVO);
 		return oAuthVO.getUserId();
